requirements.txt
Flask
flask-cors
python-pptx
Pillow
google-generativeai
groq
deepgram




